Pass segment offset to refine metadata function, update offset adjustment behaviour.
In order to calculate correct offsets when rewriting metadata, we need to know
the segment offset. This commits passes that length and updates the handling
to do so. This also addresses an edge case where if during rewriting the
varint lengths containing offsets are widened, those offsets could become
incorrect. To address this, the new metadata is recaclulated until its length
becomes stable, indicating no changes in varint sizes.